Added in alliance implementation

+++ src/main/java/io/github/tsecho/poketeams/commands/Delete.java
@@ -1,13 +1,14 @@
 package io.github.tsecho.poketeams.commands;
 
-import java.util.Map;
-
+import io.github.tsecho.poketeams.apis.AllianceAPI;
+import io.github.tsecho.poketeams.apis.PokeTeamsAPI;
+import io.github.tsecho.poketeams.enums.AllyRanks;
+import io.github.tsecho.poketeams.enums.ChatTypes;
 import io.github.tsecho.poketeams.enums.Messages.ErrorMessages;
 import io.github.tsecho.poketeams.enums.Messages.SuccessMessages;
 import io.github.tsecho.poketeams.enums.Messages.TechnicalMessages;
 import io.github.tsecho.poketeams.language.ChatUtils;
 import io.github.tsecho.poketeams.utilities.ErrorCheck;
-import io.github.tsecho.poketeams.configuration.ConfigManager;
 import io.github.tsecho.poketeams.utilities.Permissions;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.command.CommandException;
@@ -18,8 +19,7 @@ import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.command.spec.CommandSpec;
 import org.spongepowered.api.entity.living.player.Player;
 
-import io.github.tsecho.poketeams.apis.PokeTeamsAPI;
-import ninja.leaping.configurate.commented.CommentedConfigurationNode;
+import static io.github.tsecho.poketeams.configuration.ConfigManager.getStorNode;
 
 public class Delete implements CommandExecutor {
 
@@ -33,19 +33,22 @@ public class Delete implements CommandExecutor {
 
 		role = new PokeTeamsAPI(src);
 
-		if (!role.inTeam())
+		if(!role.inTeam())
 			return ErrorCheck.test(src, ErrorMessages.NOT_IN_TEAM);
-		if (!role.canDelete())
+		if(!role.canDelete())
 			return ErrorCheck.test(src, ErrorMessages.INSUFFICIENT_RANK);
 
-		for (Map.Entry<Object, ? extends CommentedConfigurationNode> teams : ConfigManager
-				.getStorNode("Teams", role.getTeam(), "Members").getChildrenMap().entrySet()) {
+		AllianceAPI alliance = new AllianceAPI(role);
 
-			ChatUtils.removeChat(teams.getKey().toString());
+		if(alliance.inAlliance() && alliance.getRank().equals(AllyRanks.OWNER.getName()))
+			return ErrorCheck.test(src, ErrorMessages.ALLY_NEEDS_LEADER);
 
-			Sponge.getServer().getPlayer(teams.getKey().toString()).ifPresent(p ->
-					p.sendMessage(SuccessMessages.DISBANDED.getText(src)));
-	    }
+		getStorNode("Teams", role.getTeam(), "Members").getChildrenMap().entrySet().stream()
+				.map(key -> key.getKey().toString())
+				.forEach(member -> {
+					ChatUtils.setChat(member, ChatTypes.PUBLIC);
+					Sponge.getServer().getPlayer(member).ifPresent(p -> p.sendMessage(SuccessMessages.DISBANDED.getText(src)));
+				});
 
 		role.deleteTeam();
